Move ticket model to its own file
